Zip containing winning entries for Code Jam 2013
Since the Code Jam 2013 on Saturday, we've received numerous requests
to share the problems and their solutions asked in the contest. As a
result, heres the set of winning entries to all the questions except
problem E for which we received no entries.

We hope this helps you. Thanks once again for being a part of Code Jam
2013.

:)

*If you are curious enough to attempt it and nail the problem, we
encourage you to share it.